🔧 Chore: loading component file move into common folder
chore #12
